+++ src/lib.rs
@@ -49,6 +49,7 @@ pub(crate) fn queue_deno_msg_or_reject_with_backpressure(
             } => deferred.reject_with_error("Runtime is closed"),
             DenoMsg::Close { deferred }
             | DenoMsg::Memory { deferred }
+            | DenoMsg::Gc { deferred }
             | DenoMsg::SetGlobal { deferred, .. }
             | DenoMsg::RegisterModule { deferred, .. }
             | DenoMsg::ClearModule { deferred, .. } => {
@@ -118,6 +119,7 @@ where
             } => deferred.reject_with_value_via_channel(mk_err("Runtime is closed")),
             DenoMsg::Close { deferred }
             | DenoMsg::Memory { deferred }
+            | DenoMsg::Gc { deferred }
             | DenoMsg::SetGlobal { deferred, .. }
             | DenoMsg::RegisterModule { deferred, .. }
             | DenoMsg::ClearModule { deferred, .. } => {